scala: note on curried functions
The there is a subtle difference between partial and curried functions. Partial
functions are multi-parameter functions that you just specify one or more
parameters first and then fill in the blanks later. Curried functions are
designed from the start to be partially filled with parameters first and
additional parameters can be specified later.